added socket reducer
