+++ src/client/reducers/gameReducer.js
@@ -1,15 +1,8 @@
-import io from 'socket.io-client';
 import * as types from '../constants/ActionTypes';
 import mockGameBoard from '../constants/mockGameBoard';
 
 const initialTeamObj = {
-  members: [
-    // {
-    //   username: '',
-    //   isSpyMaster: false,
-    //   ready: false,
-    // },
-  ],
+  members: [],
   wordsLeft: [],
 };
 
@@ -24,7 +17,7 @@ const initialState = {
     team: 'blue',
   },
   gameBoard: mockGameBoard,
-  socket: null,
+  messages: [],
 };
 
 const gameReducer = (state = initialState, action) => {
@@ -39,9 +32,6 @@ const gameReducer = (state = initialState, action) => {
         ...state,
         sessionID,
         currUser: { username, isSpyMaster: true, team: 'blue' },
-        socket: io('localhost:3000', {
-          query: `r_var=${sessionID}`,
-        }),
         blueTeam: {
           ...state.blueTeam,
           members: [...state.blueTeam.members, { username, isSpyMaster: true, ready: false }],
@@ -49,33 +39,27 @@ const gameReducer = (state = initialState, action) => {
       };
     }
     case types.JOIN_SESSION: {
-      console.log('joining session');
       const { sessionID, username } = action.payload;
-      const socket = io('localhost:3000', {
-        query: `r_var=${sessionID}`,
-      });
-      socket.emit('join session', username);
       const redCount = state.redTeam.members.length;
       const blueCount = state.redTeam.members.length;
       const team = blueCount > redCount ? 'red' : 'blue';
       const teamKey = `${team}Team`;
       const isSpyMaster = redCount === 0;
-      console.log('emitted join');
       return {
         ...state,
         sessionID,
         currUser: { username, isSpyMaster, team },
-        socket,
         [teamKey]: {
           ...state[teamKey],
           members: [...state[teamKey].members, { username, isSpyMaster, ready: false }],
         },
       };
     }
-    case types.TEST:
+    case types.NEW_MESSAGE:
       // alert(action.payload);
-      console.log('testing testing');
-      return state;
+      console.log('payload:', action.payload);
+      // delete state.socket;
+      return { ...state, messages: [...state.messages, action.payload] };
     default:
       console.log('default reducer run');
       return state;
